requirements.txt
ffmpeg-python
youtube_dl
pytgcalls==3.0.0.dev9
git+https://github.com/pyrogram/pyrogram@master
python-dotenv
tgcrypto
